sgersapp: Move experimental code to mainstream
