+++ service/static/js/main.js
@@ -1,85 +1,54 @@
-var buildings = []
-var cities = []
-var floors = ["1", "2", "3", "4", "5", "Any"];
-var times_hours = ["00", "01", "02","03",
-                   "04", "05", "06", "07",
-                   "08", "09", "10", "11",
-                   "12", "13", "14", "15",
-                   "16", "17", "18", "19",
-                   "20", "21", "22", "23"];
-var times_mins = ["00", "15", "30","45"];
-var duration_hours = ["0h", "1h", "2h","3h",
-                      "4h", "5h", "6h", "7h"];
-var duration_mins = ["00m", "15m", "30m","45m"];
-var sizes = ["1", "2", "4", "5", "6", "7", "8", "9", "10", "15", "20", "25", "30", "50", "70", "100"];
-var date_years = ["2016", "2017"];
-var date_months = ["01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"];
-var date_days = ["01", "02", "03", "04", "05", "06", "07", "08", "09", "10",
-                 "11", "12", "13", "14", "15", "16", "17", "18", "19", "20",
-                 "21", "22", "23", "24", "25", "26", "27", "28", "29", "30",
-                 "31"];
+var cities = [];
+var buildings = [];
+var floors = [];;
+var attendees = ["1", "2", "4", "5", "6", "7", "8", "9", "10", "15", "20",
+                 "25", "30", "50", "70", "100"];
+//var times = ["00:00", "01:00", "02:00", "03:00", "04:00", "05:00", "06:00",
+//             "07:00", "08:00", "09:00", "10:00", "11:00", "12:00", "13:00",
+//             "14:00", "15:00", "16:00", "17:00", "18:00", "18:00", "19:00",
+//             "20:00", "21:00", "22:00", "23:00"];
+var times = ["00:00", "00:30", "01:00", "01:30", "02:00", "02:30", "03:00",
+             "03:30", "04:00", "04:30", "05:00", "05:30", "06:00", "06:30",
+             "07:00", "07:30", "08:00", "08:30", "09:00", "09:30", "10:00",
+             "10:30", "11:00", "11:30", "12:00", "12:30", "13:00", "13:30",
+             "14:00", "14:30", "15:00", "15:30", "16:00", "16:30", "17:00",
+             "17:30", "18:00", "18:30", "19:00", "19:30", "20:00", "20:30",
+             "21:00", "21:30", "22:00", "22:30", "23:00", "23:30",];
+var startTimeIndex = null;
+var endTimeIndex = null;
 
-var DAYS_AHEAD = 0;
+var selectedRoom;
 
 function init(){
-    hideUserPassword();
     loadCitiesList();
-    createCombo(buildingSelect, buildings);
-    createCombo(citySelect, cities);
-    createCombo(floorSelect, floors);
-    createCombo(roomSizeSelect, sizes);
-    createCombo(startTimeHourSelect, times_hours);
-    createCombo(startTimeMinSelect, times_mins);
-    createCombo(durationHourSelect, duration_hours);
-    createCombo(durationMinSelect, duration_mins);
+
+    hideUsernamePasswordFields();
+    hideRoomList();
+
+    createCombo(buildingSelect,buildings);
+    createCombo(floorSelect,floors);
+    createCombo(citySelect,cities);
+    createCombo(roomSizeSelect,attendees);
 
     citySelect.value = "San Jose";
     loadBuildingList(citySelect.value);
     buildingSelect.value = "SJC19";
     loadFloorList(buildingSelect.value);
     floorSelect.value = "3";
-    durationMinSelect.value = "30m";
+
     if (navigator.geolocation) {
         navigator.geolocation.getCurrentPosition(getCity);
     }
 
-    var date = new Date();
-    var today = date.toISOString().split('T')[0];
-    bookDate.value = today;
-    var current_hour = date.getHours();
-    if (current_hour < 9) {
-        current_hour = "0" + current_hour
-    }
-    startTimeHourSelect.value = current_hour;
-}
-
-function nextDate() {
-    var nextDate = new Date();
-    DAYS_AHEAD += 1;
-    nextDate.setDate((new Date()).getDate() + DAYS_AHEAD);
-    bookDate.value = nextDate.toISOString().split('T')[0];
-}
-
-function prevDate() {
-    var nextDate = new Date();
-    if (DAYS_AHEAD > 0) {
-        DAYS_AHEAD -= 1;
-    }
-    nextDate.setDate((new Date()).getDate() + DAYS_AHEAD);
-    bookDate.value = nextDate.toISOString().split('T')[0];
-}
-
-function hideUserPassword() {
-    userLabel.style.visibility = "hidden";
-    userNameInput.style.visibility = "hidden";
-    passwordLabel.style.visibility = "hidden";
-    passwordInput.style.visibility = "hidden";
+    setTodayDate();
+    createTimeRows(times);
+    selectCurrentTime();
 }
 
 function createCombo(container, data) {
     var options = '';
     container.options.length = 0;
-    for (var i = 0; i < data.length; i++) {
+    for (var i=0; i < data.length; i++) {
         container.options.add(new Option(data[i], data[i]));
     }
 }
@@ -109,7 +78,6 @@ function loadFloorList(buildingname) {
     xmlHttp.send(null);
     floors = JSON.parse(xmlHttp.responseText);
     createCombo(floorSelect, floors);
-    resultMessage.innerHTML = "";
 }
 
 function loadBuildingList(city) {
@@ -118,30 +86,114 @@ function loadBuildingList(city) {
     xmlHttp.send(null);
     buildings = JSON.parse(xmlHttp.responseText);
     createCombo(buildingSelect, buildings);
+    resultMessage.style.display = "none";
+    resultMap.style.display = "none";
 }
 
-//Example: http://127.0.0.1:5000/showrooms?roomname=SJC19-3&starttime=2016-08-25T09:00:00&endtime=2016-08-25T19:00:00&user=mrathor&password=****
-
-
-//Example: http://127.0.0.1:5000/showrooms?roomname=SJC19-3&starttime=2016-08-25T09:00:00&endtime=2016-08-25T19:00:00&user=mrathor&password=****
-
-function submitClickHandler() {
-    var tableHeaderRowCount = 1;
-    var rowCount = mytable.rows.length;
-    for (var i = tableHeaderRowCount; i < rowCount; i++) {
-        mytable.deleteRow(tableHeaderRowCount);
-        console.log("clearing row number:" + i);
+function createTimeRows(data) {
+    var i = 0;
+    for (i = 0; i < data.length; i++) {
+        tableContainer.innerHTML += "<div class='timeTableRow'>";
+        tableContainer.innerHTML += "<label class='timeLabel' id='timeLbl" + i + "'>" + data[i]+"</label>";
+        tableContainer.innerHTML += "<button class='btn btn-default tableButton' id='timeBtn" + i + "' value='" + data[i] + "' onclick='handleTime(" + i + ")'> 30 min </button>";
+        i++;
+        tableContainer.innerHTML += "<label class='timeLabelHidden' id='timeLbl" + i + "'>" + data[i]+"</label>";
+        tableContainer.innerHTML += "<button class='btn btn-default tableButton' id='timeBtn" + i + "' value='" + data[i] + "' onclick='handleTime(" + i + ")'> 30 min </button>";
+        tableContainer.innerHTML += "</div>";
     }
-    mytable.innerHTML = "";
-    mytable.visiblity = false;
-    hideUserPassword();
 
-    var passwordb64 = encodeURIComponent(btoa(passwordInput.value));
+    tableContainer.innerHTML += "<div class='timeTableRow'>";
+    tableContainer.innerHTML += "<label class='timeLabelHidden' id='timeLbl" + i + "'" + data[0]+"</label>";
+    tableContainer.innerHTML += "<button class='btn btn-default tableButton' id='timeBtn" + i + "' value='" + data[0] + "' onclick='handleTime(" + i + ")' style='visibility:hidden'> 30 min </button>";
+    tableContainer.innerHTML += "</div>";
+}
+
+function handleTime(index) {
+    if ((startTimeIndex == null) || (index < startTimeIndex)) {
+        startTimeIndex = index;
+        endTimeIndex = index;
+    }
+    else if (startTimeIndex == endTimeIndex) {
+        endTimeIndex = index;
+    }
+    else {
+        startTimeIndex = index;
+        endTimeIndex = index;
+    }
+
+    var btn;
+    var lbl;
+    for (var i=0 ; i<times.length; i++) {
+        btn = eval('timeBtn'+i);
+        lbl = eval('timeLbl'+i);
+        if (i >= startTimeIndex && i <= endTimeIndex) {
+            btn.classList.add('tableBtnEnabled');
+            lbl.classList.add('timeLabelEnabled');
+        }
+        else {
+            btn.classList.remove('tableBtnEnabled');
+            lbl.classList.remove('timeLabelEnabled');
+        }
+    }
+}
+
+function selectCurrentTime() {
+    var date = new Date();
+    var current_hour = date.getHours();
+    var current_min = date.getMinutes();
+
+    if (current_min < 15) {
+        current_min = "00";
+    }
+    else if (current_min < 45) {
+        current_min = "30";
+    }
+    else {
+        current_hour++;
+        current_min = "00";
+    }
+
+    if (current_hour < 10) {
+        current_hour = "0" + current_hour;
+    }
+    else {
+        current_hour = "" + current_hour;
+    }
+
+    var selectTime = current_hour + ":" + current_min;
+
+    for (var i=0 ; i<times.length; i++) {
+        if (times[i] == selectTime) {
+            handleTime(i);
+            eval("timeLbl" + i).scrollIntoView(true);
+            break;
+        }
+    }
+}
+
+function setTodayDate() {
+    var today = new Date();
+    var dd = today.getDate();
+    var mm = today.getMonth() + 1; //January is 0!
+
+    var yyyy = today.getFullYear();
+    if (dd < 10) {
+        dd = '0' + dd
+    }
+    if (mm < 10) {
+        mm = '0' + mm
+    }
+    var formattedDate = yyyy +'-' + mm + '-' + dd;
+
+    dateInput.value = formattedDate;
+}
+
+function handleSearchBtnClick() {
     var timezone = new Date().getTimezoneOffset();
-
-    var queryString = `\?user=${userNameInput.value}\&password=${passwordb64}&buildingname=${buildingSelect.value}&floor=${floorSelect.value}&starttime=${bookDate.value}T${startTimeHourSelect.value}:${startTimeMinSelect.value}:00&duration=${durationHourSelect.value}${durationMinSelect.value}&attendees=${roomSizeSelect.value}&timezone=${timezone}`;
+    var starttime = (eval('timeBtn' + startTimeIndex)).value;
+    var endtime = (eval('timeBtn' + (endTimeIndex + 1))).value;
+    var queryString = `\?user=\&password=&buildingname=${buildingSelect.value}&floor=${floorSelect.value}&date=${dateInput.value}&starttime=${starttime}&endtime=${endtime}&attendees=${roomSizeSelect.value}&timezone=${timezone}`;
     loadRooms(queryString);
-
 }
 
 function loadRooms(queryString) {
@@ -157,37 +209,39 @@ function loadRooms(queryString) {
     }
 }
 
-
 function showFreeRooms(rooms_json) {
     error = rooms_json["Error"];
     if (typeof error != "undefined") {
-        mytable.innerHTML += "<td>Error: " + error + "</td>";
+        resultMessage.innerHTML = "Error: " + error;
         return;
     }
+    else {
+	resultMessage.style.display = "none";
+	resultMap.style.display = "none";
+    }
 
-    userLabel.style.visibility = "visible";
-    userNameInput.style.visibility = "visible";
-    passwordLabel.style.visibility = "visible";
-    passwordInput.style.visibility = "visible";
-    resultMessage.innerHTML = "";
-    mytable.innerHTML += "<td></td><td><small>Found " + Object.keys(rooms_json).length + " rooms</small></td>";
+    roomNamesContainer.style.display = "block";
+    roomNamesContainer.innerHTML = "<label class='rightDivLbl' id='selectRoomText'> 1. Select A Room </label>";
+    roomNamesContainer.innerHTML += "<div class='roomNamesRow'><h4>" + Object.keys(rooms_json).length + " room(s) available</h4></div>";
     for (var key in rooms_json) {
         var roomemail = rooms_json[key]["email"];
         if (typeof roomemail != "undefined") {
-            mytable.innerHTML += '<td><input type="submit" value="Reserve" onclick="bookRoom(\'' + key + '\' , \'' + roomemail + '\');"></td><td>' + key + '</td>';
-        }
-        else {
-            mytable.innerHTML += '<td></td><td>' + key + '</td>';
+            roomNamesContainer.innerHTML += "<div class='roomNamesRow'><input type='radio' name='roomRadio' value='" + key + "' onclick='handleSelectRoomBtn(this)'><label class='roomNamesRadioLbl'>" + key + "</label></div>";
         }
     }
-    mytable.visiblity = true;
 }
 
-function bookRoom(roomname, roomemail) {
+function handleSelectRoomBtn (radioBtn) {
+    selectedRoom = radioBtn.value;
+    showUsernamePasswordFields();
+}
+
+function handleReserveBtnClick() {
     var passwordb64 = encodeURIComponent(btoa(passwordInput.value));
     var timezone = new Date().getTimezoneOffset();
-
-    var queryString = `\?user=${userNameInput.value}\&password=${passwordb64}&roomname=${roomname}&roomemail=${roomemail}&starttime=${bookDate.value}T${startTimeHourSelect.value}:${startTimeMinSelect.value}:00&duration=${durationHourSelect.value}${durationMinSelect.value}&timezone=${timezone}`;
+    var starttime = (eval('timeBtn' + startTimeIndex)).value;
+    var endtime = (eval('timeBtn' + (endTimeIndex + 1))).value;
+    var queryString = `\?user=${userNameInput.value}\&password=${passwordb64}&roomname=${selectedRoom}&starttime=${starttime}&endtime=${endtime}&date=${dateInput.value}&timezone=${timezone}`;
     var xmlHttp = new XMLHttpRequest();
 
     if (userNameInput.value == "") {
@@ -213,10 +267,32 @@ function bookRoom(roomname, roomemail) {
 
     xmlHttp.open("GET", url, false); // false for synchronous request
     xmlHttp.send(null);
-    mytable.innerHTML = "";
-    mytable.visiblity = false;
-    resultMessage.innerHTML = roomname + " " + xmlHttp.responseText;
-    hideUserPassword();
+    hideRoomList();
+    hideUsernamePasswordFields();
+    resultMessage.style.display = "block";
+    resultMap.style.display = "block";
+    resultMessage.innerHTML = selectedRoom;
+    resultMap.innerHTML = xmlHttp.responseText;
 }
 
+function hideRoomList() {
+    roomNamesContainer.innerHTML = '';
+    roomNamesContainer.style.display = "none";
+}
+
+function showUsernamePasswordFields() {
+    var userpassHTML = "<label class='rightDivLbl'> 2. Enter User name and Password </label>";
+    userpassHTML += "<div><label class='userNamePswdLbl'>User name</label></div>";
+    userpassHTML += "<input  type='text' id='userNameInput' class='form-control userNamePswdInputText' placeholder='User name'></input>";
+    userpassHTML += "<div><label class='userNamePswdLbl'>Password</label></div>";
+    userpassHTML += "<input  type='password' id='passwordInput' class='form-control userNamePswdInputText' placeholder='Password'></input>";
+    userpassHTML += "<button class='btn btn-default reserveButton form-control' type='button' onclick='handleReserveBtnClick()' >Reserve</button>";
+
+    usernamePasswordContainer.style.display = "block";
+    usernamePasswordContainer.innerHTML = userpassHTML;
+}
+
+function hideUsernamePasswordFields() {
+    usernamePasswordContainer.style.display = "none";
+}
