--- service/templates/index2.html
@@ -1,147 +1,0 @@
-<html>
-
-<head>
-    <script type="text/javascript" src="static/js/main.js"></script>
-    <style>
-        body {
-            background-color: white;
-            color: black;
-            font-family: "Verdana";
-        }
-        #contact{
-            color: black;
-            font-size: 10px;
-            text-align: right;
-        }
-        #searchpane{
-            background-color: #FFFFFF;
-            color: black;
-            font-size: 15px;
-            text-align: left;
-            float: left;
-            width: 30%;
-            height: 100%;
-        }
-        #banner{
-            background-color: #AAAAAA;
-            color: white;
-            font-size: 60px;
-            padding: 10px;
-            border-left: 10px solid orange;
-            border-radius: 5px;
-            text-align: center;
-        }
-        #resultpane{
-            background-color: #EFEFEF;
-            color: black;
-            font-size: 15px;
-            text-align: center;
-            float: right;
-            width: 70%;
-            height: 100%;
-        }
-        label{
-            display: inline-block;
-            text-align: left;
-            margin-bottom: 3px;
-        }
-        input {
-            width: 130px;
-            height: 25px;
-        }
-        #userLabel{
-            width: 130px;
-        }
-        #submitButton {
-            margin-top: 30px;
-            height: 30px;
-            width: 90%;
-            font-family: "Verdana";
-            font-size: 16px;
-            font-weight: bold;
-        }
-        fieldset{
-            border: none;
-            width: 650px;
-            margin: 0px auto;
-        }
-        #mytable{
-            margin: auto;
-            margin-top: 30px;
-            width: 650px;
-        }
-        td{
-            padding: 5px;
-        }
-    </style>
-</head>
-<body onload="init()">
-<div id="optionContainer">
-    <div id="banner">Conference Room Finder</div>
-    <div id="contact">
-        <a href="mailto:roomfinder@cisco.com">Report a bug</a> &nbsp; &nbsp; &nbsp;
-        <a href="mailto:sgersapp@cisco.com">Contact Us</a> &nbsp; &nbsp; &nbsp;
-    </div>
-</div>
-
-<br/>
-<div id="searchpane">
-    <table width="100%">
-    <tr>
-        <td>
-            <label id="cityLabel">City</label> <br/>
-            <select id="citySelect" onchange="loadBuildingList(this.value)"></select>
-        </td>
-
-        <td align="left">
-            <label id="roomSizeLabel">Attendees</label> <br/>
-            <select id="roomSizeSelect"></select>
-        </td>
-    </tr>
-    <tr>
-        <td>
-            <label id="buildingLabel">Building</label> <br />
-            <select id="buildingSelect" onchange="loadFloorList(this.value)"></select><br/>
-        </td>
-
-        <td>
-            <label id="floorLabel">Floor</label> <br/>
-            <select id="floorSelect"></select>
-        </td>
-    </tr>
-    <tr>
-        <td colspan="2">
-            <label id="dateLabel">Date</label> <br/>
-            <button id="prev" onclick="prevDate()">&lt;</button>
-            <input type="date" style="width:60%" id="bookDate"/>
-            <button id="next" onclick="nextDate()">&gt;</button>
-        </td>
-    </tr>
-    <tr>
-        <td>
-            <label id="startTimeLabel">Time</label> <br/>
-            <select id="startTimeHourSelect"></select>
-            <select id="startTimeMinSelect"></select>
-        </td>
-        <td>
-            <label id="durationLabel">Duration</label> <br />
-            <select id="durationHourSelect"></select>
-            <select id="durationMinSelect"></select>
-        </td>
-    </tr>
-    </table>
-
-    <button id="submitButton" width="100%" onclick="submitClickHandler()">Search</button><br/><br/>
-</div>
-
-
-<div id="resultpane">
-
-    <label id="userLabel">Username</label><input type="text" id="userNameInput"></input>
-    <label id="passwordLabel">Password</label><input type="password" id="passwordInput"></input>
-    <table id="mytable">
-    </table>
-</div>
-
-</body>
-</html>
