--- service/templates/index3.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Room finder</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-
-    <script type="text/javascript" src="static/js/main3.js"></script>
-    <link rel="stylesheet" href="static/css/style.css">
-</head>
-<body onload="init()">
-
-<div>
-    <div class="titleContainer">
-        <h2 class="titleText">Conference Room Finder</h2>
-    </div>
-    <div class="leftAndRightDivContainer">
-        <div class="form-group tableRow">
-            <div class="leftDiv">
-                <div>
-                    <label class="cityLabel">City</label>
-                    <label class="attendeesLabel">Attendees</label>
-                </div>
-                <div>
-                    <select id="citySelect" class="form-control selectStyle" onchange="loadBuildingList(this.value)"></select>
-                    <select id="roomSizeSelect" class="form-control selectStyle"></select>
-                </div>
-                <div>
-                    <label class="buildingLabel">Building</label>
-                    <label class="floorLabel">Floor</label>
-                </div>
-                <div class="buildingFloorSelectContainer">
-                    <select id="buildingSelect" class="form-control selectStyle" onchange="loadFloorList(this.value)"></select>
-                    <select id="floorSelect" class="form-control selectStyle"></select>
-                </div>
-                <div class="dateContainer">
-                    <label class="dateText">Date</label>
-                    <input id="dateInput" type="date" class="form-control dateInput" />
-                </div>
-                <label class="timeText">Time</label>
-                <hr>
-                <div id="tableContainer" class="tableContainer" > </div>
-
-                <button class="btn btn-default searchButton form-control" type="button" onclick="handleSearchBtnClick()" >Search</button>
-            </div>
-
-            <div class="rightDiv">
-                <div id="resultMessage" class="roomNamesRow" ></div>
-	    	<div id="resultMap" class="mapContainer" ></div>
-                <div id="roomNamesContainer"></div>
-                <div id="usernamePasswordContainer"></div>
-                <!-- <div><h4 class="userNamePswdLbl">Site Map</h4></div> -->
-            </div>
-        </div>
-    </div>
-</div>
-
-</body>
-</html>
